VexUser.c code
This was the main code we were using the day f he competition for
jenison. I know that it is rather bloated and full of stuff that needs
to be removed. If you code help us identify anything that we are doing
horrible wrong or something that we could be doing better.